feat: OK

+++ src/FileUploadAspNetCore/Models/FileInfo.cs
@@ -2,7 +2,7 @@ namespace FileUploadAspNetCore.Models
 {
     public class FileInfo
     {
-        public int Size { get; set; }
+        public long Size { get; set; }
         public string Name { get; set; }
         public string Path { get; set; }
         public int Id { get; set; }
